+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 **Internal**:
 
+- Introduce a new profile format called `sample`. ([#1462](https://github.com/getsentry/relay/pull/1462))
 - Generate a new profile ID when splitting a profile for multiple transactions. ([#1473](https://github.com/getsentry/relay/pull/1473))
 - Pin Rust version to 1.63.0 in Dockerfile. ([#1482](https://github.com/getsentry/relay/pull/1482))
 - Normalize measurement units in event payload. ([#1488](https://github.com/getsentry/relay/pull/1488))